requirements.txt
pydantic~=1.10.2
PyYAML~=6.0
pytest~=7.1.3
devtools~=0.9.0
fastapi~=0.85.0